a few adjustments were made.

+++ core/data/src/main/java/com/app/data/model/FoodModel.kt
@@ -9,10 +9,10 @@ import android.os.Parcelable
 data class FoodModel(
     @DocumentId val id: String? = null,
     val label: String = "",
-    val calories: Double = 0.0,
-    val protein: Double = 0.0,
-    val fat: Double = 0.0,
-    val carbohydrates: Double = 0.0,
+    val calories: Int = 0,
+    val protein: Int = 0,
+    val fat: Int = 0,
+    val carbohydrates: Int = 0,
     val image: String = "",
     val date: String = "",
     val year: String = "",
@@ -21,15 +21,14 @@ data class FoodModel(
     val dayName: String = ""
 ) : Serializable
 
-
 @Parcelize
 data class FoodModelParcelize(
     @DocumentId val id: String? = null,
     val label: String = "",
-    val calories: Double = 0.0,
-    val protein: Double = 0.0,
-    val fat: Double = 0.0,
-    val carbohydrates: Double = 0.0,
+    val calories: Int = 0,
+    val protein: Int = 0,
+    val fat: Int = 0,
+    val carbohydrates: Int = 0,
     val image: String = "",
     val date: String = "",
     val year: String = "",
